% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_greenhouse_jobs.R
\name{get_greenhouse_jobs}
\alias{get_greenhouse_jobs}
\title{Get All Greenhouse Jobs}
\usage{
get_greenhouse_jobs(greenhouse_companies)
}
\arguments{
\item{greenhouse_companies}{A tibble with columns `id`, `company`, and `tw`, which stand for
the Greenhouse ID, company name, and twitter handle respectively.}
}
\description{
Get All Greenhouse Jobs
}
