% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lever_jobs.R
\name{get_lever_jobs}
\alias{get_lever_jobs}
\title{Get All Lever Jobs}
\usage{
get_lever_jobs(lever_companies)
}
\arguments{
\item{lever_companies}{A tibble with columns `id`, `company`, and `tw`, which stand for
the Lever ID, company name, and twitter handle respectively.}
}
\description{
Get All Lever Jobs
}
