% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_greenhouse_jobs.R
\name{greenhouse_api_jobs}
\alias{greenhouse_api_jobs}
\title{Get Greenhouse API Jobs}
\usage{
greenhouse_api_jobs(board_token)
}
\arguments{
\item{board_token}{The Greenhouse ID of the company to query jobs.}
}
\description{
Get Greenhouse API Jobs
}
