% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lever_jobs.R
\name{lever_api_jobs}
\alias{lever_api_jobs}
\title{Get Lever API Jobs}
\usage{
lever_api_jobs(board_token)
}
\arguments{
\item{board_token}{The Lever ID of the company to query jobs.}
}
\description{
Get Lever API Jobs
}
